+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 3.1.0a7 (unreleased)
 --------------------
 
+Feature:
+
+- Added coresandbox fixture for Volto's cypress tests
+  [sneridagh]
+
 Internal:
 
 - Test with Plone 6.0.0a2